refactor(open-mpm): split tools/ batch 3 — native_ticketing/memory/native_search/native_memory (closes #361)
Split four oversized tool modules into focused submodule directories:

- native_ticketing.rs (1310) -> native_ticketing/{crud,workflow,actions,mod,tests}.rs
- memory.rs (1162) -> memory/{recall,vector_search,mod,tests}.rs
- native_search.rs (1086) -> native_search/{helpers,search_code,search_memory,search_skills,mod}.rs
- native_memory.rs (980) -> native_memory/{backend,store_memory,retrieve_memory,list_keys,mod,tests}.rs

No public API surface changes; MemoryBackend remains pub at
crate::tools::native_memory::MemoryBackend. All non-test source files are
under the 500-line cap; large shared-mock test blocks extracted to tests.rs.
